+++ .dir-locals.el
@@ -1,3 +1,3 @@
 ((emacs-lisp-mode . ((indent-tabs-mode . nil)
-                     (fill-column . 100)
+                     (fill-column . 80)
                      (elisp-lint-indent-specs . ((when-let . 1))))))